+++ sources/packages/backend/apps/api/src/route-controllers/application-offering-change-request/models/application-offering-change-request.institutions.dto.ts
@@ -1,0 +1,10 @@
+/**
+ * Eligible applications offering change list details.
+ */
+export class ApplicationOfferingChangeSummaryAPIOutDTO {
+  applicationNumber: string;
+  applicationId: number;
+  studyStartPeriod: string;
+  studyEndPeriod: string;
+  fullName: string;
+}
