+++ sources/packages/backend/apps/api/src/route-controllers/index.ts
@@ -67,3 +67,4 @@ export * from "./student-appeal/student-appeal.institutions.controller";
 export * from "./restriction/restriction.institutions.controller";
 export * from "./restriction/restriction.controller.service";
 export * from "./note/note.institutions.controller";
+export * from "./application-offering-change-request/application-offering-change-request.institutions.controller";
