+++ sources/packages/backend/apps/api/src/services/index.ts
@@ -39,3 +39,4 @@ export * from "./student-account-applications/student-account-applications.model
 export * from "./education-program-offering/education-program-offering-import-csv.service";
 export * from "./education-program-offering/education-program-offering-validation.models";
 export * from "./education-program-offering/education-program-offering-validation.service";
+export * from "./application-offering-change-request/application-offering-change-request.service";
