+++ sources/packages/backend/apps/db-migrations/src/sql/Types/Add-application-offering-change-assessment-trigger-type.sql
@@ -1,3 +1,3 @@
 ALTER TYPE sims.assessment_trigger_types
 ADD
-  VALUE if not exists 'Application offering change';
+  VALUE 'Application offering change';
