+++ sources/packages/backend/libs/sims-db/src/index.ts
@@ -4,3 +4,4 @@ export * from "./data.model.service";
 export * from "./error-handler";
 export * from "./data-source";
 export * from "./database-utils";
+export * from "./pagination-utils";
