+++ sources/packages/web/src/components/layouts/Institution/sidebar/HomeSideBar.vue
@@ -125,7 +125,7 @@ export default defineComponent({
                 props: {
                   prependIcon: "fa:far fa-folder",
                   to: {
-                    name: InstitutionRoutesConst.REQUEST_A_CHANGE,
+                    name: InstitutionRoutesConst.REQUEST_CHANGE,
                     params: {
                       locationId: data.id,
                     },
