+++ sources/packages/web/src/services/ApplicationOfferingChangeRequestService.ts
@@ -1,0 +1,35 @@
+import { PaginationOptions } from "../types";
+import ApiClient from "./http/ApiClient";
+import {
+  PaginatedResultsAPIOutDTO,
+  ApplicationOfferingChangeSummaryAPIOutDTO,
+} from "@/services/http/dto";
+
+export class ApplicationOfferingChangeRequestService {
+  // Share Instance
+  private static instance: ApplicationOfferingChangeRequestService;
+
+  static get shared(): ApplicationOfferingChangeRequestService {
+    return this.instance || (this.instance = new this());
+  }
+
+  /**
+   * Gets all eligible applications that can be requested for application
+   * offering change.
+   * @param locationId location id.
+   * @param paginationOptions options to execute the pagination.
+   * @returns list of eligible applications that can be requested for
+   * application offering change.
+   */
+  async getEligibleApplicationOfferingChangeRecords(
+    locationId: number,
+    paginationOptions: PaginationOptions,
+  ): Promise<
+    PaginatedResultsAPIOutDTO<ApplicationOfferingChangeSummaryAPIOutDTO>
+  > {
+    return ApiClient.ApplicationOfferingChangeRequestApi.getEligibleApplicationOfferingChangeRecords(
+      locationId,
+      paginationOptions,
+    );
+  }
+}
