+++ sources/packages/web/src/services/InstitutionService.ts
@@ -16,8 +16,6 @@ import {
   AESTCreateInstitutionAPIInDTO,
   PrimaryIdentifierAPIOutDTO,
   OptionItemAPIOutDTO,
-  PaginatedResultsAPIOutDTO,
-  ApplicationOfferingChangeSummaryAPIOutDTO,
 } from "@/services/http/dto";
 
 export class InstitutionService {
@@ -160,25 +158,5 @@ export class InstitutionService {
    */
   async getMyInstitutionLocationsDetails() {
     return ApiClient.InstitutionLocation.getMyInstitutionLocationsDetails();
-  }
-
-  /**
-   * Gets all eligible application that can be requested for application
-   * offering change.
-   * @param locationId location id.
-   * @param paginationOptions options to execute the pagination.
-   * @returns list of eligible application that can be requested for
-   * application offering change.
-   */
-  async getEligibleApplicationOfferingChangeRecords(
-    locationId: number,
-    paginationOptions: PaginationOptions,
-  ): Promise<
-    PaginatedResultsAPIOutDTO<ApplicationOfferingChangeSummaryAPIOutDTO>
-  > {
-    return ApiClient.Institution.getEligibleApplicationOfferingChangeRecords(
-      locationId,
-      paginationOptions,
-    );
   }
 }
