+++ sources/packages/web/src/services/http/ApplicationOfferingChangeRequestApi.ts
@@ -1,0 +1,30 @@
+import HttpBaseClient from "./common/HttpBaseClient";
+import { PaginationOptions } from "@/types";
+import {
+  PaginatedResultsAPIOutDTO,
+  ApplicationOfferingChangeSummaryAPIOutDTO,
+} from "@/services/http/dto";
+import { getPaginationQueryString } from "@/helpers";
+
+export class ApplicationOfferingChangeRequestApi extends HttpBaseClient {
+  /**
+   * Gets all eligible applications that can be requested for application
+   * offering change.
+   * @param locationId location id.
+   * @param paginationOptions options to execute the pagination.
+   * @returns list of eligible applications that can be requested for
+   * application offering change.
+   */
+  async getEligibleApplicationOfferingChangeRecords(
+    locationId: number,
+    paginationOptions: PaginationOptions,
+  ): Promise<
+    PaginatedResultsAPIOutDTO<ApplicationOfferingChangeSummaryAPIOutDTO>
+  > {
+    let url = `location/${locationId}/application-offering-change-request/available?`;
+    url += getPaginationQueryString(paginationOptions, true);
+    return this.getCall<
+      PaginatedResultsAPIOutDTO<ApplicationOfferingChangeSummaryAPIOutDTO>
+    >(this.addClientRoot(url));
+  }
+}
