+++ sources/packages/web/src/services/http/dto/ApplicationOfferingChangeRequest.dto.ts
@@ -1,0 +1,10 @@
+/**
+ * Eligible applications offering change list details.
+ */
+export interface ApplicationOfferingChangeSummaryAPIOutDTO {
+  applicationNumber: string;
+  applicationId: number;
+  studyStartPeriod: string;
+  studyEndPeriod: string;
+  fullName: string;
+}
