+++ sources/packages/web/src/types/AppRoutes.ts
@@ -36,7 +36,7 @@ export enum AppRoutes {
   LocationProgramsOfferingsCreate = "location-offerings/location/:locationId/education-program/:programId/create",
   LocationUsers = "location-users/",
   ActiveApplicationsSummary = "active-applications/location/:locationId/summary",
-  RequestAChange = "request-change/location/:locationId/summary",
+  RequestApplicationOfferingChange = "location/:locationId/application-offering-change",
   ActiveApplicationEdit = "active-applications/location/:locationId/application/:applicationId",
   LocationProgramInfoRequestSummary = "program-info-request/location/:locationId/summary",
   LocationCOESummary = "confirmation-of-enrollment/location/:locationId/summary",
